add mock_setup to test_utils
