+++ agent/src/tools/liquidity_request.py
@@ -1,5 +1,4 @@
 import json
-import os
 import requests
 
 from decimal import Decimal
